Created a helper function for category Put for api
