+++ Helper/api/categoryHelper.js
@@ -77,3 +77,54 @@ export const Get = async (res) => {
     }
   });
 };
+
+// update;
+export const Put = async (body, res) => {
+  await prisma.category
+    .update({
+      where: { id: body.id },
+      data: { title: body.title, description: body.content },
+    })
+    .then(async () => {
+      const goParams = {
+        Bucket: keys.aws.s3Bucket,
+        Key: `category/${body.title}.jpeg`,
+      };
+
+      const base64Data = new Buffer.from(
+        body.image.replace(/^data:image\/\w+;base64,/, ''),
+        'base64'
+      );
+      const params = {
+        Bucket: keys.aws.s3Bucket,
+        Key: `category/${body.title}.jpeg`,
+        Body: base64Data,
+        ACL: 'public-read',
+        ContentEncoding: 'base64',
+        ContentType: `image/jpeg`,
+      };
+      s3.deleteObject(goParams, async (err) => {
+        if (err) {
+          console.log('err', err);
+        } else {
+          await s3
+            .upload(params)
+            .promise()
+            .then(async (re) => {
+              const { Key, Location } = re;
+              await prisma.category
+                .update({
+                  where: { id: body.id },
+                  data: { s3BucketKey: Key, image: Location },
+                })
+                .then(() => {
+                  res.json({
+                    status: 200,
+                    message: 'Data successfully updated',
+                  });
+                });
+            });
+        }
+      });
+    });
+};
